chru_dot_edu/requirements.txt
biopython (1.65)
Django (1.8.1)
django-admin-bootstrapped (2.5.0)
django-bootstrap3 (5.4.0)
django-tinymce (1.5.3)
django-twitter-bootstrap (3.3.0)
pip (7.0.1)
setuptools (15.0)
xmltodict (0.9.2)
